moved MsgDisplay to context
